YOLO_detection/requirements.txt
# YOLO Detection Requirements for RF Signal Detection

# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
Pillow>=9.5.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# YOLO specific
pycocotools>=2.0.6
lap>=0.4.0
thop>=0.1.1
